crear causa

+++ services/causaService.js
@@ -15,12 +15,15 @@ class CausaService {
 
   async updateCausa(id, causaData) {
     const updatedCausa = await causaRepository.updateCausa(id, causaData);
+    if (!updatedCausa) throw new Error('Causa no encontrada');
     return new CausaDTO(updatedCausa);
   }
 
   async deleteCausa(id) {
+    const causa = await causaRepository.getCausaById(id);
+    if (!causa) throw new Error('Causa no encontrada');
     await causaRepository.deleteCausa(id);
-  } 
+  }
 
   async getTotalCausasByDay() {
     const totalCausasByDay = await causaRepository.getTotalCausasByDay();
@@ -29,12 +32,9 @@ class CausaService {
 
   async getCausasByUserId(userId) {
     const causas = await causaRepository.getCausasByUserId(userId);
+    if (!causas || causas.length === 0) throw new Error('No se encontraron causas para el usuario');
     return causas.map(causa => new CausaDTO(causa));
   }
 }
 
 module.exports = new CausaService();
-
-
-
-
